# PERF: Improve performance of SiteSetting.humanize_name + all_settings